chore(scap): move plugin_info.h to engine/source_plugin
