+++ userspace/libscap/engine/source_plugin/source_plugin_public.h
@@ -13,7 +13,7 @@ limitations under the License.
 
 #pragma once
 
-#include "../../plugin_info.h"
+#include "plugin_info.h"
 
 #define SOURCE_PLUGIN_ENGINE "source_plugin"
